useProfileRecord 로 유저 레이팅 받아옴 useContextMenus 로 메뉴 정보들 분리
LogoutButton 삭제. logoutAction 을 ContextMenu 폴더로 이동
안쓰는 메뉴항목들 좀 삭제. 리팩토링
