--- front/app/(main)/@home/(nav)/LogoutButton.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { SquareButton } from "@components/Button/SquareButton";
-import { logoutAction } from "./logoutAction";
-
-export function LogoutButton() {
-    return <SquareButton onClick={() => logoutAction()}>Log out</SquareButton>;
-}
