+++ front/app/__libs__/components/ContextMenu/ContextMenu.tsx
@@ -1,167 +1,16 @@
 import { ContextMenuBase } from "./ContextMenuBase";
 import { useAtomValue } from "jotai";
 import { TargetedAccountUUIDAtom } from "@atoms/AccountAtom";
-import { useProtectedProfile } from "@hooks/useProfile";
+import { useProfileRecord, useProtectedProfile } from "@hooks/useProfile";
 import { useEffect, useRef, useState } from "react";
-import { RoleNumber } from "@common/generated/types";
 import { useChatMember } from "@hooks/useChatRoom";
 import {
     useCurrentAccountUUID,
     useCurrentChatRoomUUID,
 } from "@hooks/useCurrent";
 import { useContextMenuActions } from "./useContextMenuActions";
-
-export type Relationship = "myself" | "friend" | "stranger";
-
-export type Scope = "ChatRoom" | "FriendModal" | "Navigation";
-
-type ProfileMenuActions =
-    | "copytag"
-    | "changeactivestatus"
-    | "addfriend"
-    | "directmessage"
-    | "modifyfriend"
-    | "logout"
-    | "deletefriend"
-    | "gotoprofile"
-    | "addenemy"
-    | "accessban"
-    | "sendban"
-    | "reportuser"
-    | "transfer"
-    | "changerole";
-
-type ProfileMenu = {
-    name: string;
-    description?: string | undefined;
-    action?: ProfileMenuActions | undefined;
-    relation: Relationship[];
-    isImportant: boolean;
-    minRoleLevel?: number | undefined;
-    scope: Scope[];
-    className: string;
-    UI?: React.ReactNode | undefined;
-};
-
-const profileMenus: ProfileMenu[] = [
-    {
-        name: "태그 복사",
-        action: "copytag",
-        relation: ["myself", "friend", "stranger"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom"],
-        className: "",
-    },
-    {
-        name: "내 상태 변경",
-        action: "changeactivestatus",
-        relation: ["myself"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["Navigation"],
-        className: "",
-    },
-    {
-        name: "로그아웃",
-        action: "logout",
-        relation: ["myself"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["Navigation"],
-        className: "",
-    },
-    {
-        name: "상세 프로필",
-        action: "gotoprofile",
-        relation: ["myself", "friend", "stranger"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom", "FriendModal", "Navigation"],
-        className: "",
-    },
-    {
-        name: "다이렉트 메시지",
-        action: "directmessage",
-        relation: ["friend"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom", "FriendModal"],
-        className: "",
-    },
-    {
-        name: "친구 추가",
-        action: "addfriend",
-        relation: ["stranger"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom"],
-        className: "",
-    },
-    {
-        name: "친구 삭제",
-        action: "deletefriend",
-        relation: ["friend"],
-        isImportant: true,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom", "FriendModal", "Navigation"],
-        className: "hover:bg-red-500/30",
-    },
-    {
-        name: "신고",
-        action: "reportuser",
-        relation: ["friend", "stranger"],
-        isImportant: true,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom"],
-        className: "hover:bg-red-500/30",
-    },
-    {
-        name: "차단",
-        action: "addenemy",
-        relation: ["friend", "stranger"],
-        isImportant: false,
-        minRoleLevel: undefined,
-        scope: ["ChatRoom", "FriendModal", "Navigation"],
-        className: "hover:bg-tertiary/30",
-    },
-    {
-        name: "채팅 금지",
-        action: "sendban",
-        relation: ["friend", "stranger"],
-        isImportant: true,
-        minRoleLevel: RoleNumber.MANAGER,
-        scope: ["ChatRoom"],
-        className: "hover:bg-tertiary/30",
-    },
-    {
-        name: "추방",
-        action: "accessban",
-        relation: ["friend", "stranger"],
-        isImportant: true,
-        minRoleLevel: RoleNumber.MANAGER,
-        scope: ["ChatRoom"],
-        className: "hover:bg-tertiary/30",
-    },
-    {
-        name: "소유권 양도",
-        action: "transfer",
-        relation: ["friend", "stranger"],
-        minRoleLevel: RoleNumber.ADMINISTRATOR,
-        isImportant: true,
-        scope: ["ChatRoom"],
-        className: "hover:bg-red-500/30",
-    },
-    {
-        name: "매니저 임명/해임",
-        action: "changerole",
-        relation: ["friend", "stranger"],
-        minRoleLevel: RoleNumber.ADMINISTRATOR,
-        isImportant: true,
-        scope: ["ChatRoom"],
-        className: "hover:bg-red-500/30",
-    },
-];
+import type { ProfileMenu, Scope } from "./useContextMenus";
+import { useContextMenus } from "./useContextMenus";
 
 function ContextMenuItem({
     menuInfo,
@@ -227,8 +76,13 @@ export function ContextMenu({ type }: { type: Scope }) {
     const profile = useProtectedProfile(targetAccountUUID);
     const currentChatRoomUUID = useCurrentChatRoomUUID();
     const currentAccountUUID = useCurrentAccountUUID();
+
     const currentUser = useChatMember(currentChatRoomUUID, currentAccountUUID);
     const roleLevel = Number(currentUser?.role ?? 0);
+
+    const record = useProfileRecord(targetAccountUUID);
+
+    const menus = useContextMenus();
     const actions = useContextMenuActions(
         targetAccountUUID,
         currentAccountUUID,
@@ -243,10 +97,10 @@ export function ContextMenu({ type }: { type: Scope }) {
             ? "friend"
             : "stranger";
 
-    //TODO: fetch score
-    const score = 1321;
+    const skillRating = record?.skillRating ?? "-";
     const rating: ProfileMenu = {
-        name: `rating: ${score}`,
+        name: `rating: ${skillRating}`,
+        action: "noaction",
         relation: ["myself", "friend", "stranger"],
         scope: ["ChatRoom"],
         isImportant: false,
@@ -254,8 +108,8 @@ export function ContextMenu({ type }: { type: Scope }) {
     };
 
     return (
-        <ContextMenuBase className="w-full">
-            {[rating, ...profileMenus]
+        <ContextMenuBase>
+            {[rating, ...menus]
                 .filter((menu) => {
                     return (
                         menu.scope.includes(type) &&
@@ -268,11 +122,7 @@ export function ContextMenu({ type }: { type: Scope }) {
                         <ContextMenuItem
                             key={menu.name}
                             menuInfo={menu}
-                            action={
-                                menu.action !== undefined
-                                    ? actions[menu.action]
-                                    : undefined
-                            }
+                            action={actions[menu.action]}
                         />
                     );
                 })}
