+++ front/app/__libs__/components/ContextMenu/index.ts
@@ -1,2 +1,2 @@
-export type { Relationship } from "./ContextMenu";
+export type { Relationship, Scope, ProfileMenu } from "./useContextMenus";
 export { ContextMenu } from "./ContextMenu";
