+++ front/app/__libs__/components/ProfileItem/ProfileItem.tsx
@@ -2,8 +2,8 @@ import { TargetedAccountUUIDAtom } from "@atoms/AccountAtom";
 import { Avatar } from "@components/Avatar";
 import { useProtectedProfile, usePublicProfile } from "@hooks/useProfile";
 import { Provider, createStore } from "jotai";
+import type { Scope } from "@components/ContextMenu";
 import { ContextMenu } from "@components/ContextMenu";
-import type { Scope } from "@components/ContextMenu/ContextMenu";
 import type { AccountProfilePublicPayload } from "@common/profile-payloads";
 import { Icon } from "@components/ImageLibrary";
 import { useChatMember } from "@hooks/useChatRoom";
